added cogs in cogs folder
